Enchancement wallet sync status design changed in chat
