+++ ts/components/BchatMainPanel.tsx
@@ -1,12 +1,14 @@
 import React, { useEffect } from 'react';
-import {  useSelector } from 'react-redux';
+import {  useDispatch, useSelector } from 'react-redux';
 import { useAppIsFocused } from '../hooks/useAppFocused';
 import { getFocusedSettingsSection } from '../state/selectors/section';
 import { SmartBchatConversation } from '../state/smart/BchatConversation';
 import { BchatSettingsView } from './settings/BchatSettings';
 import { getOurPubKeyStrFromCache } from '../bchat/utils/User';
 import { getConversationById } from '../data/data';
-import { updateBchatUpgradeInstructionModal  } from '../state/ducks/modalDialog';
+import { updateBchatUpgradeInstructionModal, updateBchatWalletPasswordModal  } from '../state/ducks/modalDialog';
+import { SettingsKey } from '../data/settings-key';
+import { getWalletSyncBarShowInChat } from '../state/selectors/walletConfig';
 
 const FilteredSettingsView = BchatSettingsView as any;
 
@@ -21,11 +23,30 @@ export async function getconverstation() {
 export const BchatMainPanel = () => {
   const focusedSettingsSection = useSelector(getFocusedSettingsSection);
   const isSettingsView = focusedSettingsSection !== undefined;
+  const dispatch = useDispatch();
+  const chatwithWallet = window.getSettingValue(SettingsKey.settingsChatWithWallet) || false;
+  // const WalletSyncInitiatedWithChat=useSelector(getWalletSyncInitiatedWithChat)
+ const walletSyncBarShowInChat=useSelector(getWalletSyncBarShowInChat);
   
   useEffect(() => {
     getconverstation();
+    validation()
 
-  }, []);
+  }, [chatwithWallet]);
+  
+  const walletPassWordValidation = () => {
+    // console.log('WalletSyncInitiatedWithChat ::',WalletSyncInitiatedWithChat);
+    
+    if (chatwithWallet && !walletSyncBarShowInChat) {
+
+      dispatch(updateBchatWalletPasswordModal({}))
+    }
+
+  }
+  const validation = () => {
+    walletPassWordValidation();
+    // chatInstruction && chatWithWalletInstruction();
+  }
   // even if it looks like this does nothing, this does update the redux store.
   useAppIsFocused();
   if (isSettingsView) {
